requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
anthropic>=0.25.0
python-dotenv>=1.0.0
openai>=1.0.0